requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.4
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker==7.1.0
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h11==0.8.1
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
joblib==1.4.2
json5==0.9.25
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
lean==1.0.196
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maskpass==0.3.7
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
multidict==6.0.5
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pynput==1.7.7
pyopenweather==0.1.1
pyparsing==3.1.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
quantconnect==0.1.0
quantconnect-stubs==16470
requests==2.32.3
requests-async==0.6.2
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
srt==3.5.3
stack-data==0.6.3
TA-Lib==0.4.30
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vosk==0.3.45
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
